requirements.txt
anthropic==0.3.5
colorama==0.4.6
Pillow==10.0.0
pygments==2.15.1
python-dotenv==1.0.0
tavily-python==0.1.9
tiktoken==0.4.0
openai==0.27.0
groq==0.3.2
qdrant-client==1.1.1
sentence-transformers==2.2.2
pydantic==1.10.7